requirements.txt
asgiref==3.7.2
Django==4.2.7
django-browser-reload==1.12.1
paho-mqtt==1.6.1
sqlparse==0.4.4
tzdata==2023.3
